% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k3k2.R
\name{k3k2}
\alias{k3k2}
\title{Create a k3k2 object}
\usage{
k3k2(samples)
}
\arguments{
\item{samples}{A numeric vector, or a list of numeric vectors}
}
\value{
An object of class 'k3k2'
}
\description{
Create a k3k2 object
}
