% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k3k2Diagram.R
\name{k3k2Diagram}
\alias{k3k2Diagram}
\title{Create a k3k2Diagram object}
\usage{
k3k2Diagram(data = NULL, reference = c("Gamma", "Gamma Inverse"))
}
\arguments{
\item{data}{A k3k2 object (optional)}

\item{reference}{Character vector of reference curves; default c("Gamma", "Gamma Inverse")}

\item{design_parameters}{List of parameters (point size, colours...)}
}
\value{
An object of class 'k3k2Diagram'
}
\description{
Create a k3k2Diagram object
}
