% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logCumulants_wrapper.R
\name{logcumulants_wrapper}
\alias{logcumulants_wrapper}
\title{Compute log-cumulants (tilde kappa2 and kappa3)}
\usage{
logcumulants_wrapper(x)
}
\arguments{
\item{x}{Numeric vector of positive values (no NA)}
}
\value{
Named numeric vector with elements 'kappa2' and 'kappa3'
}
\description{
Compute log-cumulants (tilde kappa2 and kappa3)
}
