% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k3k2Diagram.R
\name{plot_k3k2}
\alias{plot_k3k2}
\title{Plot method for k3k2Diagram objects}
\usage{
plot_k3k2(
  x,
  line_size = 1,
  point_size = 3,
  text_size = 12,
  colours = c("blue", "red"),
  ...
)
}
\description{
Plot method for k3k2Diagram objects
}
